+++ classes/client.js
@@ -1,21 +1,51 @@
 const Discord = require("discord.js");
-const { sep } = require("path");
+const path = require("path");
 const now = require("performance-now");
-const CommandMessage = require("./classes/commandMessage.js");
-const Loader = require("./classes/loader.js");
-const ArgResolver = require("./classes/argResolver.js");
+const CommandMessage = require("./commandMessage.js");
+const Loader = require("./loader.js");
+const ArgResolver = require("./argResolver.js");
+const PermLevels = require("./permLevels.js");
  /* Will Change this later */
-const Config = require("./classes/Configuration Types/Config.js");
+const Config = require("./Configuration Types/Config.js");
 
-require("./classes/Extendables.js");
+require("./Extendables.js");
 
-/* eslint-disable no-throw-literal, no-use-before-define, no-restricted-syntax */
+const defaultPermStructure = new PermLevels()
+  .addLevel(0, false, () => true)
+  .addLevel(2, false, (client, msg) => {
+    if (!msg.guild) return false;
+    const modRole = msg.guild.roles.find("name", msg.guild.conf.modRole);
+    return modRole && msg.member.roles.has(modRole.id);
+  })
+  .addLevel(3, false, (client, msg) => {
+    if (!msg.guild) return false;
+    const adminRole = msg.guild.roles.find("name", msg.guild.conf.adminRole);
+    return adminRole && msg.member.roles.has(adminRole.id);
+  })
+  .addLevel(4, false, (client, msg) => {
+    if (!msg.guild) return false;
+    return msg.author.id === msg.guild.owner.id;
+  })
+  .addLevel(9, true, (client, msg) => msg.author.id === client.config.ownerID)
+  .addLevel(10, false, (client, msg) => msg.author.id === client.config.ownerID);
+
+/* eslint-disable no-throw-literal, no-use-before-define, no-restricted-syntax, no-underscore-dangle */
 module.exports = class Komada extends Discord.Client {
 
   constructor(config = {}) {
     if (typeof config !== "object") throw new TypeError("Configuration for Komada must be an object.");
     super(config.clientOptions);
     this.config = config;
+    this.config.disabled = config.disabled || {};
+    this.config.disabled = {
+      commands: config.disabled.commands || [],
+      events: config.disabled.events || [],
+      functions: config.disabled.functions || [],
+      inhibitors: config.disabled.inhibitors || [],
+      finalizers: config.disabled.finalizers || [],
+      monitors: config.disabled.monitors || [],
+      providers: config.disabled.providers || [],
+    };
     this.funcs = new Loader(this);
     this.argResolver = new ArgResolver(this);
     this.helpStructure = new Map();
@@ -40,23 +70,25 @@ module.exports = class Komada extends Discord.Client {
       escapeMarkdown: Discord.escapeMarkdown,
       splitMessage: Discord.splitMessage,
     };
-    this.coreBaseDir = `${__dirname}${sep}`;
-    this.clientBaseDir = `${process.env.clientDir || process.cwd()}${sep}`;
-    this.outBaseDir = `${process.env.outDir || process.cwd()}${sep}`;
+    this.coreBaseDir = path.join(__dirname, "../");
+    this.clientBaseDir = `${process.env.clientDir || process.cwd()}${path.sep}`;
+    this.outBaseDir = `${process.env.outDir || process.cwd()}${path.sep}`;
     this.compiledLangs = process.env.compiledLangs || process.env.compiledLang || [];
     if (!Array.isArray(this.compiledLang)) this.compiledLangs = [this.compiledLangs];
     this.guildConfs = Config.guildConfs;
     this.configuration = Config;
     this.application = null;
+    this.once("ready", this._ready.bind(this));
   }
 
   get invite() {
+    if (!this.user.bot) throw "Why would you need an invite link for a selfbot...";
     const permissions = Discord.Permissions.resolve([...new Set(this.commands.reduce((a, b) => a.concat(b.conf.botPerms), ["READ_MESSAGES", "SEND_MESSAGES"]))]);
     return `https://discordapp.com/oauth2/authorize?client_id=${this.application.id}&permissions=${permissions}&scope=bot`;
   }
 
   validatePermStructure() {
-    const permStructure = this.config.permStructure || defaultPermStructure;
+    const permStructure = this.config.permStructure || defaultPermStructure.structure;
     if (!Array.isArray(permStructure)) throw "PermStructure must be an array.";
     if (permStructure.some(perm => typeof perm !== "object" || typeof perm.check !== "function" || typeof perm.break !== "boolean")) {
       throw "Perms must be an object with a check function and a break boolean.";
@@ -67,43 +99,40 @@ module.exports = class Komada extends Discord.Client {
 
   async login(token) {
     const start = now();
-    await this.loadEverything();
+    await this.funcs.loadAll(this);
     this.emit("log", `Loaded in ${(now() - start).toFixed(2)}ms.`);
     super.login(token);
   }
 
-  async loadEverything() {
-    await this.funcs.loadAll(this);
-    this.once("ready", async () => {
-      this.config.prefixMention = new RegExp(`^<@!?${this.user.id}>`);
-      this.application = await super.fetchApplication();
-      await Promise.all(Object.keys(this.funcs).map((key) => {
-        if (this.funcs[key].init) return this.funcs[key].init(this);
-        return true;
-      }));
-      await Promise.all(this.providers.map((piece) => {
-        if (piece.init) return piece.init(this);
-        return true;
-      }));
-      await Promise.all(this.commands.map((piece) => {
-        if (piece.init) return piece.init(this);
-        return true;
-      }));
-      await Promise.all(this.commandInhibitors.map((piece) => {
-        if (piece.init) return piece.init(this);
-        return true;
-      }));
-      await Promise.all(this.commandFinalizers.map((piece) => {
-        if (piece.init) return piece.init(this);
-        return true;
-      }));
-      await Promise.all(this.messageMonitors.map((piece) => {
-        if (piece.init) return piece.init(this);
-        return true;
-      }));
-      await this.configuration.initialize(this);
-      this.ready = true;
-    });
+  async _ready() {
+    this.config.prefixMention = new RegExp(`^<@!?${this.user.id}>`);
+    if (this.user.bot) this.application = await super.fetchApplication();
+    await Promise.all(Object.keys(this.funcs).map((key) => {
+      if (this.funcs[key].init) return this.funcs[key].init(this);
+      return true;
+    }));
+    await Promise.all(this.providers.map((piece) => {
+      if (piece.init) return piece.init(this);
+      return true;
+    }));
+    await Promise.all(this.commands.map((piece) => {
+      if (piece.init) return piece.init(this);
+      return true;
+    }));
+    await Promise.all(this.commandInhibitors.map((piece) => {
+      if (piece.init) return piece.init(this);
+      return true;
+    }));
+    await Promise.all(this.commandFinalizers.map((piece) => {
+      if (piece.init) return piece.init(this);
+      return true;
+    }));
+    await Promise.all(this.messageMonitors.map((piece) => {
+      if (piece.init) return piece.init(this);
+      return true;
+    }));
+    await this.configuration.initialize(this);
+    this.ready = true;
   }
 
   sweepCommandMessages(lifetime = this.commandMessageLifetime) {
@@ -126,74 +155,6 @@ module.exports = class Komada extends Discord.Client {
   }
 
 };
-
-const defaultPermStructure = [
-  {
-    check: () => true,
-    break: false,
-  },
-  {
-    check: () => false,
-    break: false,
-  },
-  {
-    check: (client, msg) => {
-      if (!msg.guild) return false;
-      const modRole = msg.guild.roles.find("name", msg.guild.conf.modRole);
-      if (modRole && msg.member.roles.has(modRole.id)) return true;
-      return false;
-    },
-    break: false,
-  },
-  {
-    check: (client, msg) => {
-      if (!msg.guild) return false;
-      const adminRole = msg.guild.roles.find("name", msg.guild.conf.adminRole);
-      if (adminRole && msg.member.roles.has(adminRole.id)) return true;
-      return false;
-    },
-    break: false,
-  },
-  {
-    check: (client, msg) => {
-      if (!msg.guild) return false;
-      if (msg.author.id === msg.guild.owner.id) return true;
-      return false;
-    },
-    break: false,
-  },
-  {
-    check: () => false,
-    break: false,
-  },
-  {
-    check: () => false,
-    break: false,
-  },
-  {
-    check: () => false,
-    break: false,
-  },
-  {
-    check: () => false,
-    break: false,
-  },
-  {
-    check: (client, msg) => {
-      if (msg.author.id === client.config.ownerID) return true;
-      return false;
-    },
-    break: true,
-  },
-  {
-    check: (client, msg) => {
-      if (msg.author.id === client.config.ownerID) return true;
-      return false;
-    },
-    break: false,
-  },
-];
-
 
 process.on("unhandledRejection", (err) => {
   if (!err) return;