+++ index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  Client: require('./classes/client'),
+  PermLevels: require('./classes/permLevels')
+};